move tests to gc.rs
